CAF-3518 - JS Post Processing
